# Speedup activerecord tests