@zzak 2024-02-24
@fatkodima @byroot How about to re-enable debug log on CI? I also wondered if we had the facility or wanted to retrieve coredumps in some cases.