@rafaelfranca 2024-02-26
He means this change https://github.com/rails/rails/pull/49030/files#diff-1080ad0e6eb23991c5d3ac5a71adc545f7977d47da5df19079e0502b2ed09628R26

We need a new environment variable to reanable it if we want to on CI.